auth: move getHomeDir to only be included on Windows as well
