+++ auth_sha1_windows.go
@@ -7,6 +7,7 @@ import (
 	"crypto/sha1"
 	"encoding/hex"
 	"os"
+	"os/user"
 )
 
 // AuthCookieSha1 returns an Auth that authenticates as the given user with the
@@ -99,4 +100,23 @@ func (a authCookieSha1) generateChallenge() []byte {
 	enc := make([]byte, 32)
 	hex.Encode(enc, b)
 	return enc
+}
+
+// Get returns the home directory of the current user, which is usually the
+// value of HOME environment variable. In case it is not set or empty, os/user
+// package is used.
+//
+// If linking statically with cgo enabled against glibc, make sure the
+// osusergo build tag is used.
+//
+// If needing to do nss lookups, do not disable cgo or set osusergo.
+func getHomeDir() string {
+	homeDir := os.Getenv("HOME")
+	if homeDir != "" {
+		return homeDir
+	}
+	if u, err := user.Current(); err == nil {
+		return u.HomeDir
+	}
+	return "/"
 }
